+++ test/fixtures/searchRuntimeDependencies/customProbe.js
@@ -1,0 +1,2 @@
+const danger = 'danger';
+const stream = eval('require')('stream');
